cid#1506514 Uninitialized scalar field
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/136099
Tested-by: Jenkins
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/138507
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>

+++ starmath/inc/ElementsDockingWindow.hxx
@@ -39,7 +39,7 @@ class SmElementsControl
 
     SmDocShell*   mpDocShell;
     SmFormat      maFormat;
-    size_t        mnCurrentSetIndex;
+    int           mnCurrentSetIndex;
     sal_uInt16    m_nSmSyntaxVersion;
 
     bool          mbVerticalMode;
@@ -49,7 +49,7 @@ class SmElementsControl
     Link<OUString, void> maSelectHdlLink;
 
     void addElement(const OUString& aElementVisual, const OUString& aElementSource, const OUString& aHelpText);
-    void addElements(size_t nCategory);
+    void addElements(int nCategory);
 
     void build();
 
@@ -65,7 +65,7 @@ public:
     ~SmElementsControl();
 
     static const std::vector<TranslateId>& categories();
-    void setElementSetIndex(size_t nSetIndex);
+    void setElementSetIndex(int nSetIndex);
 
     void setVerticalMode(bool bVertical);
